requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-types==0.19.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djangorestframework-types==0.8.0
psycopg2==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
types-psycopg2==2.9.21.20240417
